+++ server.go
@@ -1,3 +1,7 @@
+// Copyright 2024 HUMAN. All rights reserved.
+// Use of this source code is governed by a MIT style
+// license that can be found in the LICENSE file.
+
 package envite
 
 import (
@@ -7,6 +11,7 @@ import (
 	"net/http"
 )
 
+// Server is an HTTP server, serving UI and API requests to manage the Environment when running in ExecutionModeDaemon.
 type Server struct {
 	addr       string
 	env        *Environment
@@ -14,6 +19,7 @@ type Server struct {
 	errHandler func(string)
 }
 
+// NewServer creates a new Server instance for the given Environment.
 func NewServer(port string, env *Environment) *Server {
 	if len(port) > 0 && port[0] != ':' {
 		port = ":" + port
@@ -28,6 +34,7 @@ func NewServer(port string, env *Environment) *Server {
 	return s
 }
 
+// Start starts the HTTP server.
 func (s *Server) Start() error {
 	err := s.httpServer.ListenAndServe()
 	if err != nil && !errors.Is(err, http.ErrServerClosed) {
@@ -36,6 +43,7 @@ func (s *Server) Start() error {
 	return nil
 }
 
+// Close gracefully shuts down the HTTP server.
 func (s *Server) Close() error {
 	s.httpServer.SetKeepAlivesEnabled(false)
 	return s.httpServer.Shutdown(context.Background())