+++ ui/src/SplitPanel.tsx
@@ -1,3 +1,7 @@
+// Copyright 2024 HUMAN. All rights reserved.
+// Use of this source code is governed by a MIT style
+// license that can be found in the LICENSE file.
+
 import React, { useEffect, useState } from 'react';
 import './SplitPanel.css';
 import { Message, Status } from './api';