+++ CHANGELOG.md
@@ -4,3 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [[0.0.1](https://github.com/PerimeterX/envite/releases/tag/v0.0.1)] - 2023-12-21
+
+Initial release.
+
+## [[0.0.2](https://github.com/PerimeterX/envite/compare/v0.0.1...v0.0.2)] - 2023-12-24
+
+### Fixed
+
+- fix error handling on stop container.
+
+## [[0.0.3](https://github.com/PerimeterX/envite/compare/v0.0.2...v0.0.3)] - 2024-02-17
+
+Prepare for open source.
+
+### Added
+
+- CLI support.
+- Go releaser support.
+- Unit tests.
+- Go docs.
+- Improved README.md.
+- Added several small changes to the SDK API to allow smoother experience in the CLI.